webdacs/DoAn_AntiSpam/requirements.txt
streamlit
scikit-learn
numpy
pandas
joblib
pytesseract
Pillow
Flask-Cors
